Add two-level persistence to keyedReducer
Previously the `keyedReducer` was "eating" the SERIALIZE and DESERIALIZE
actions because they never had the attached `key` value. This meant that
the entire state was being persisted (I think)

Now they pass through those actions to each item in the collection on
SERIALIZE and return through an optional serializer and deserializer.